organize component widgets structure
