+++ lib/pages/assignment/assignment_check_page/widgets/close_button_widget.dart
