+++ lib/pages/assignment/assignment_list_page/widgets/assignment_card.dart
@@ -1,0 +1,142 @@
+import 'package:flutter/material.dart';
+import 'package:school_erp/pages/assignment/assignment_check_page/assignment_check_page.dart';
+
+class AssignmentCard extends StatelessWidget {
+  final String subject;
+  final String title;
+  final String assignDate;
+  final String lastSubmissionDate;
+  final String status;
+  final Color statusColor;
+
+  const AssignmentCard({
+    super.key,
+    required this.subject,
+    required this.title,
+    required this.assignDate,
+    required this.lastSubmissionDate,
+    required this.status,
+    required this.statusColor,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: () {
+        Navigator.push(
+          context,
+          PageRouteBuilder(
+            pageBuilder: (context, animation, secondaryAnimation) =>
+                const AssignmentCheckPage(),
+            transitionsBuilder:
+                (context, animation, secondaryAnimation, child) {
+              const begin = 0.0;
+              const end = 1.0;
+              const curve = Curves.easeInOut;
+              var fadeTween =
+                  Tween(begin: begin, end: end).chain(CurveTween(curve: curve));
+              var fadeAnimation = animation.drive(fadeTween);
+
+              return Stack(
+                children: [
+                  Container(
+                    color: Colors.transparent,
+                  ),
+                  FadeTransition(
+                    opacity: fadeAnimation,
+                    child: child,
+                  ),
+                ],
+              );
+            },
+          ),
+        );
+      },
+      child: Card(
+        margin: const EdgeInsets.symmetric(vertical: 8.0),
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(10.0),
+        ),
+        child: Stack(
+          children: [
+            Padding(
+              padding: const EdgeInsets.all(16.0),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Container(
+                    padding: const EdgeInsets.symmetric(
+                        vertical: 4.0, horizontal: 8.0),
+                    decoration: BoxDecoration(
+                      color: Colors.blue.shade100,
+                      borderRadius: BorderRadius.circular(5.0),
+                    ),
+                    child: Text(
+                      subject,
+                      style: const TextStyle(
+                        color: Color(0xFF5278C1),
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                  ),
+                  const SizedBox(height: 8.0),
+                  Text(
+                    title,
+                    style: const TextStyle(
+                      fontSize: 16.0,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                  const SizedBox(height: 4.0),
+                  Text(
+                    'Assign Date: $assignDate',
+                    style: const TextStyle(
+                      color: Colors.grey,
+                      fontSize: 14.0,
+                    ),
+                  ),
+                  Text(
+                    'Last Submission Date: $lastSubmissionDate',
+                    style: const TextStyle(
+                      color: Colors.grey,
+                      fontSize: 14.0,
+                    ),
+                  ),
+                  const SizedBox(height: 8.0),
+                  ElevatedButton(
+                    onPressed: () {
+                      // Handle status button action here
+                    },
+                    style: ElevatedButton.styleFrom(
+                      backgroundColor: statusColor,
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(20.0),
+                      ),
+                    ),
+                    child: Text(
+                      status,
+                      style: const TextStyle(
+                        color: Colors.white,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            Positioned(
+              top: 8.0,
+              right: 8.0,
+              child: IconButton(
+                icon: const Icon(Icons.edit, color: Colors.grey),
+                onPressed: () {
+                  // Handle edit button press
+                  print('Edit button pressed');
+                },
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
